[R] refactor to dynamic commands
